requirements.txt
langchain==0.1.11
matplotlib==3.8.3
python-dotenv==1.0.1
streamlit==1.32.0
openai==0.28.0